nala: Rework PRIMITIVE_TYPES check (#9)
The old check wasn't appropriate as the order of qualifiers isn't
enforced by pycparser nor C compilers, yet Nala enforced the order.

This is a problem with stdatomic.h which contains instances of
"short unsigned int", a type not recognized by Nala.

Instead just check that all qualifiers and the core types are common,
leaving the qualifier validation to pycarser and the compiler.
Utilize a set for this.

Encapsulate the PRIMITIVE_TYPES list in a method and call that instead
of implementing the set check everywhere.
This results in some changes to is_primitive_type_pointer and
is_primitive_type_pointer_type to no longer take a types argument,
instead they call the new is_primitive_type method as necessary.

Co-authored-by: John Ernberg <john.ernberg@actia.se>